+++ UMC.WApp/frmShift.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using DevExpress.XtraBars;
+using DevExpress.XtraEditors;
+using System;
 using System.Data;
 using System.Data.Entity.Migrations;
 using System.Linq;
@@ -24,26 +26,35 @@ namespace UMC.WApp
         {
             try
             {
-                if(!string.IsNullOrEmpty(txtName.Text))
+                string promptValue = ShowDialog("Please Input Code", "Confirm Code");
+                if(promptValue == "umcvn@123321")
                 {
-                    ShiftViewModel shiftVm = new ShiftViewModel();
-                    shiftVm.Name = txtName.Text;
-                    shiftVm.StartHour = int.Parse(cbbStartHour.SelectedItem.ToString());
-                    shiftVm.StartMinute = int.Parse(cbbStartMinute.SelectedItem.ToString());
-                    shiftVm.EndHour = int.Parse(cbbEndHour.SelectedItem.ToString());
-                    shiftVm.EndMinute = int.Parse(cbbEndMinute.SelectedItem.ToString());
+                    if (!string.IsNullOrEmpty(txtName.Text))
+                    {
+                        ShiftViewModel shiftVm = new ShiftViewModel();
+                        shiftVm.Name = txtName.Text;
+                        shiftVm.StartHour = int.Parse(cbbStartHour.SelectedItem.ToString());
+                        shiftVm.StartMinute = int.Parse(cbbStartMinute.SelectedItem.ToString());
+                        shiftVm.EndHour = int.Parse(cbbEndHour.SelectedItem.ToString());
+                        shiftVm.EndMinute = int.Parse(cbbEndMinute.SelectedItem.ToString());
 
-                    Shift newShift = new Shift();
-                    newShift.UpdateShift(shiftVm);
-                    db.Shifts.Add(newShift);
-                    db.SaveChanges();
-                    ClearData();
-                    LoadData();
-                    MessageBox.Show("Thêm mới thành công!");
+                        Shift newShift = new Shift();
+                        newShift.UpdateShift(shiftVm);
+                        db.Shifts.Add(newShift);
+                        db.SaveChanges();
+                        ClearData();
+                        LoadData();
+                        MessageBox.Show("Thêm mới thành công!");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Bạn không được để trống dữ liệu", "Error",
+                                 MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
                 }
-                else
+               else
                 {
-                    MessageBox.Show("Bạn không được để trống dữ liệu", "Error",
+                    MessageBox.Show("Bạn không được thêm mới", "Error",
                              MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
             }
@@ -68,35 +79,45 @@ namespace UMC.WApp
         {
             try
             {
-                if (txtName.Text != "")
+                string promptValue = ShowDialog("Please Input Code", "Confirm Code");
+                if (promptValue == "umcvn@123321")
                 {
-                    ShiftViewModel shiftVm = new ShiftViewModel();
-                    shiftVm.Name = txtName.Text;
-                    shiftVm.StartHour = int.Parse(cbbStartHour.Text.ToString());
-                    shiftVm.StartMinute = int.Parse(cbbStartMinute.Text.ToString());
-                    shiftVm.EndHour = int.Parse(cbbEndHour.Text.ToString());
-                    shiftVm.EndMinute = int.Parse(cbbEndMinute.Text.ToString());
+                    if (txtName.Text != "")
+                    {
+                        ShiftViewModel shiftVm = new ShiftViewModel();
+                        shiftVm.Name = txtName.Text;
+                        shiftVm.StartHour = int.Parse(cbbStartHour.Text.ToString());
+                        shiftVm.StartMinute = int.Parse(cbbStartMinute.Text.ToString());
+                        shiftVm.EndHour = int.Parse(cbbEndHour.Text.ToString());
+                        shiftVm.EndMinute = int.Parse(cbbEndMinute.Text.ToString());
 
-                    var id = Convert.ToInt32(dgvShift.Rows[dgvShift.CurrentRow.Index].Cells[0].Value);
-                    shiftVm.ID = id;
+                        var id = Convert.ToInt32(dgvShift.Rows[dgvShift.CurrentRow.Index].Cells[0].Value);
+                        shiftVm.ID = id;
 
-                    Shift newShift = new Shift();
-                    newShift.UpdateShift(shiftVm);
+                        Shift newShift = new Shift();
+                        newShift.UpdateShift(shiftVm);
 
-                    db.Set<Shift>().AddOrUpdate(newShift);
-                    db.SaveChanges();
+                        db.Set<Shift>().AddOrUpdate(newShift);
+                        db.SaveChanges();
 
-                    ClearData();
-                    LoadData();
-                    MessageBox.Show("Cập nhật thành công!");
+                        ClearData();
+                        LoadData();
+                        MessageBox.Show("Cập nhật thành công!");
+                    }
+                    else
+                    {
+                        MessageBox.Show("Please Select Record to Update", "Error",
+                            MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
                 }
                 else
                 {
-                    MessageBox.Show("Please Select Record to Update","Error",
-                        MessageBoxButtons.OK,MessageBoxIcon.Error);
+                    MessageBox.Show("Bạn không được Update", "Error",
+                              MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
+                
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 MessageBox.Show("Cập nhật không thành công !, Error:" + ex.Message);
             }
@@ -118,22 +139,53 @@ namespace UMC.WApp
             return db.Shifts.Find(id);
         }
 
+        public static string ShowDialog(string text, string caption)
+        {
+            Form prompt = new Form()
+            {
+                Width = 500,
+                Height = 150,
+                FormBorderStyle = FormBorderStyle.FixedDialog,
+                Text = caption,
+                StartPosition = FormStartPosition.CenterScreen
+            };
+            Label textLabel = new Label() { Left = 50, Top = 20, Text = text };
+            TextBox textBox = new TextBox() { Left = 50, Top = 50, Width = 400, UseSystemPasswordChar = true };
+            Button confirmation = new Button() { Text = "Ok", Left = 350, Width = 100, Top = 70, DialogResult = DialogResult.OK };
+            confirmation.Click += (sender, e) => { prompt.Close(); };
+            prompt.Controls.Add(textBox);
+            prompt.Controls.Add(confirmation);
+            prompt.Controls.Add(textLabel);
+            prompt.AcceptButton = confirmation;
+
+            return prompt.ShowDialog() == DialogResult.OK ? textBox.Text : "";
+        }
+
         private void btnDelete_Click(object sender, EventArgs e)
         {
             try
             {
-                var id = Convert.ToInt32(dgvShift.Rows[dgvShift.CurrentRow.Index].Cells[0].Value);
-                var shift = db.Shifts.Find(id);
-                db.Shifts.Remove(shift);
-                db.SaveChanges();
-                MessageBox.Show("Đã xóa bản ghi !");
-                ClearData();
-                LoadData();
+                string promptValue = ShowDialog("Please Input Code", "Confirm Code");
+                if (promptValue == "umcvn@123321")
+                {
+                    var id = Convert.ToInt32(dgvShift.Rows[dgvShift.CurrentRow.Index].Cells[0].Value);
+                    var shift = db.Shifts.Find(id);
+                    db.Shifts.Remove(shift);
+                    db.SaveChanges();
+                    MessageBox.Show("Đã xóa bản ghi !");
+                    ClearData();
+                    LoadData();
+                }
+                else
+                {
+                    MessageBox.Show("Error", "Error",
+                                  MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 MessageBox.Show("Error: Không xóa được bản ghi !" + ex.Message);
-            }
+            }          
         }
         public void ClearData()
         {
